requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# CrewAI and Agent Framework
crewai==0.28.8
langchain==0.1.9
langchain-openai==0.0.5
langchain-anthropic==0.1.1

# LLM Providers
openai==1.12.0
anthropic==0.18.0

# ElevenLabs
elevenlabs==1.0.0

# GitHub Analysis
PyGithub==2.2.0
GitPython==3.1.41

# PowerPoint Analysis
python-pptx==0.6.23
Pillow==10.2.0

# Code Analysis (optional)
radon==6.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Data Processing
requests==2.31.0
aiohttp==3.9.3
